# refactor: use arena to allocate docs to further improve performance