… Cargo.toml
… examples/pretty_test.rs
… src/pretty/mod.rs
… src/pretty/mode.rs
… src/pretty/parened_expr.rs
… src/pretty/snapshots/typstyle_core__pretty__items__tests__pretty_items_fit-2.snap
… src/pretty/snapshots/typstyle_core__pretty__items__tests__pretty_items_fit-3.snap
… src/pretty/snapshots/typstyle_core__pretty__items__tests__pretty_items_fit.snap
… src/pretty/style.rs
… src/pretty/table.rs
… src/pretty/util.rs
… src/snapshots/typstyle_core__util__tests__pretty_items_single-2.snap
… src/snapshots/typstyle_core__util__tests__pretty_items_single-3.snap
… src/snapshots/typstyle_core__util__tests__pretty_items_single.snap
… src/util.rs