This PR uses `Arena` in pretty to allocate docs, which can greatly improve performance in terms of both time and space. 

Compared with previous results, the time consumption is expected to decrease by 50% to 75%.

```
pretty-tablex           time:   [4.6388 ms 4.7914 ms 4.9576 ms]
                        change: [-72.193% -70.728% -69.297%] (p = 0.00 < 0.05)

pretty-cetz-manual      time:   [1.2813 ms 1.3052 ms 1.3370 ms]
                        change: [-53.188% -51.617% -50.093%] (p = 0.00 < 0.05)

pretty-undergraduate-math
                        time:   [1.3302 ms 1.3477 ms 1.3647 ms]
                        change: [-54.371% -52.070% -49.875%] (p = 0.00 < 0.05)

pretty-packages/codly   time:   [2.2947 ms 2.3554 ms 2.4220 ms]
                        change: [-77.193% -75.772% -74.509%] (p = 0.00 < 0.05)

pretty-packages/fletcher-diagram
                        time:   [662.12 µs 688.35 µs 719.36 µs]
                        change: [-75.143% -74.252% -73.242%] (p = 0.00 < 0.05)

pretty-packages/fletcher-draw
                        time:   [1.4852 ms 1.5006 ms 1.5160 ms]
                        change: [-76.428% -74.798% -72.106%] (p = 0.00 < 0.05)

pretty-packages/touying/core
                        time:   [2.8626 ms 2.9032 ms 2.9493 ms]
                        change: [-78.116% -77.676% -77.199%] (p = 0.00 < 0.05)

pretty-packages/touying/utils
                        time:   [1.5441 ms 1.5625 ms 1.5814 ms]
                        change: [-71.776% -71.282% -70.763%] (p = 0.00 < 0.05)
```